pyproject.toml
[project]
name = "2025-fastapi-seminar-template-2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "argon2-cffi>=25.1.0",
    "fastapi[standard]>=0.116.1",
    "passlib[bcrypt]>=1.7.4",
    "bcrypt>=3.2,<4.0",
    "python-jose[cryptography]>=3.5.0",
    "pytest>=8.4.2",
    "freezegun>=1.5.5",
]
